Extract convert logic into hook
