+++ src/scss/_buttons.scss
@@ -31,7 +31,7 @@
     --webkit-transition: all 300ms ease;
   }
 
-  &:hover {
+  &:hover:not(:disabled) {
     box-shadow: 0 0.5rem 1rem rgba(136, 195, 117, 0.2);
     transition: all 300ms ease;
     --webkit-transition: all 300ms ease;
@@ -41,6 +41,10 @@
       transition: all 300ms ease;
       --webkit-transition: all 300ms ease;
     }
+  }
+  
+  &:disabled {
+    cursor: not-allowed;
   }
 
   span {
