+++ src/scss/_recipe.scss
@@ -53,8 +53,7 @@
   display: flex;
   gap: 0.75rem;
   font-size: 1.5rem;
-  -webkit-user-select: none;
-          user-select: none;
+  user-select: none;
   cursor: pointer;
 }
 
@@ -83,8 +82,7 @@
   border-bottom: 2px solid $color-primary;
   text-align: center;
   cursor: pointer;
-  -webkit-user-select: none;
-          user-select: none;
+  user-select: none;
 }
 
 .conversion-dropdown-grams {
